pyproject.toml
[project]
name = "quarto-coursegen"
version = "0.1.0"
description = "Scaffold Quarto course websites from a course.yaml definition"
readme = "README.md"
authors = [
    { name = "Sebastian Cacean"}
]
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "jinja2>=3.1",
    "pyyaml>=6",
]

[project.scripts]
quarto-coursegen = "quarto_coursegen.cli:app"

[dependency-groups]
dev = [
    "pytest>=8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/quarto_coursegen"]
